+++ app/templates/sources/src/test/groovy/package/AbstractKitTest.groovy
@@ -3,9 +3,8 @@ package <%= projectPackage %>
 import org.apache.tools.ant.taskdefs.condition.Os
 import org.gradle.testkit.runner.BuildResult
 import org.gradle.testkit.runner.GradleRunner
-import org.junit.Rule
-import org.junit.rules.TemporaryFolder
 import spock.lang.Specification
+import spock.lang.TempDir
 
 /**
  * Base class for Gradle TestKit based tests.
@@ -19,12 +18,11 @@ abstract class AbstractKitTest extends Specification {
     boolean debug
     boolean isWin = Os.isFamily(Os.FAMILY_WINDOWS)
 
-    @Rule
-    final TemporaryFolder testProjectDir = new TemporaryFolder()
+    @TempDir File testProjectDir
     File buildFile
 
     def setup() {
-        buildFile = testProjectDir.newFile('build.gradle')
+        buildFile = file('build.gradle')
         // jacoco coverage support
         fileFromClasspath('gradle.properties', 'testkit-gradle.properties')
     }
@@ -34,7 +32,7 @@ abstract class AbstractKitTest extends Specification {
     }
 
     File file(String path) {
-        new File(testProjectDir.root, path)
+        new File(testProjectDir, path)
     }
 
     File fileFromClasspath(String toFile, String source) {
@@ -52,7 +50,7 @@ abstract class AbstractKitTest extends Specification {
     }
 
     String projectName() {
-        return testProjectDir.root.getName()
+        return testProjectDir.getName()
     }
 
     GradleRunner gradle(File root, String... commands) {
@@ -65,7 +63,7 @@ abstract class AbstractKitTest extends Specification {
     }
 
     GradleRunner gradle(String... commands) {
-        gradle(testProjectDir.root, commands)
+        gradle(testProjectDir, commands)
     }
 
     BuildResult run(String... commands) {
@@ -89,6 +87,6 @@ abstract class AbstractKitTest extends Specification {
     protected String unifyString(String input) {
         return input
                 // cleanup win line break for simpler comparisons
-                .replaceAll("\r", '')
+                .replace("\r", '')
     }
 }