+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 * Disable gradle metadata publishing
 * Remove java version selection: java8 used by default for compilation
 * Enable mavenLocal as plugin repository by default (in settings.gradle)
+* settings.gradle is modified now on project update
 
 ### 1.9.0 (2020-01-26)
 * Update ru.vyarus.java-lib 1.1.2 -> 2.1.0